requirements.txt
asgiref==3.2.7
Django==3.0.6
djangorestframework==3.11.0
mysqlclient==1.4.6
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
